% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{basisFuns}
\alias{basisFuns}
\title{De Boors algorithm to find basis functions}
\usage{
basisFuns(x, degree, knots)
}
\arguments{
\item{x}{`double` Point to search for position in knots.}

\item{degree}{`unsigned int` Degree of the polynomial between the knots.}

\item{knots}{`arma::vec` Vector of knots. It's the users responsibility to
pass a **SORTED** vector.}
}
\value{
`arma::sp_mat` Sparse matrix containing the basis functions.
}
\description{
De Boors algorithm to find basis functions
}
