% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binnedMatMultResponse}
\alias{binnedMatMultResponse}
\title{Calculating binned matrix product for response term}
\usage{
binnedMatMultResponse(X, y, k, w)
}
\arguments{
\item{X}{[\code{arma::mat}]\cr
Matrix X.}

\item{y}{[\code{arma::vec}]\cr
Response vector y.}

\item{k}{[\code{arma::uvec}]\cr
Index vector for mapping to original matrix $X_o(i,) = X(k(i),.)$.}

\item{w}{[\code{arma::vec}]\cr
Vector of weights that are accumulated.}
}
\value{
\code{arma::mat} Matrix Product $X^TWX$.
}
\description{
This function calculates the matrix product using Algorithm 3 of Zheyuan Li, Simon N. Wood: "Faster
model matrix crossproducts for large generalized linear models with discretized covariates". The idea
is to compute just on the unique rows of X by also using an index vector to map to the original matrix.
The algorithm implemented here is a small adaption of the original algorithm. Instead of calculating $XW$
which again, needs to be transposed, we directly calculate $X^TW$ to avoid another transposing step. In addition
to the original algorithm the algorithm here directly calculates the crossproduct with the response.
}
\examples{
nsim = 1e6L
nunique = trunc(sqrt(nsim))

xunique = runif(n = nunique, min = 0, max = 10)
k = sample(x = seq_len(nunique), size = nsim, replace = TRUE)

X = poly(x = xunique, degree = 20L)
y = runif(nsim)

binnedMatMultResponse(X = X, y = y, k = k-1, w = 1)
}
