% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{binVectorCustom}
\alias{binVectorCustom}
\title{Calculate vector of bins of specific size}
\usage{
binVectorCustom(x, n_bins)
}
\arguments{
\item{x}{[\code{arma::vec}]\cr
Vector that should be discretized.}

\item{n_bins}{[\code{unsigned int}]\cr
Number of unique points for binning the vector x.}
}
\value{
\code{arma::vec} Vector of discretized x.
}
\description{
This function returns a vector of equally spaced points of length n_bins.
}
\examples{
\dontrun{
x = runif(100)
binVectorCustom(x, 10)
}
}
