% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateIndexVector}
\alias{calculateIndexVector}
\title{Calculate index vector for binned vector}
\usage{
calculateIndexVector(x, x_bins)
}
\arguments{
\item{x}{[\code{arma::vec}]\cr
Vector that should be discretized.}

\item{x_bins}{[\code{arma::vec}]\cr
Vector of unique values for binning.}
}
\value{
\code{arma::uvec} Index vector.
}
\description{
This function returns the indexes of the unique values to the complete binned vector.
}
\examples{
x = runif(100)
bins = binVector(x)
idx = calculateIndexVector(x, bins)
data.frame(x = x, bins = bins[idx])
}
