% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{centerMatrix}
\alias{centerMatrix}
\title{Subtract the basis of one matrix of another}
\usage{
centerMatrix(X1, X2)
}
\arguments{
\item{X1}{[`matrix()`]\cr
Matrix of rank p1 from which the X1 basis is subtracted.}

\item{X2}{[`matrix()`]\cr
Matrix of rank p2.}
}
\value{
`matrix()` Matrix of rank p1-p2 with new basis
}
\description{
This function subtracts the basis of X2 from X1. The resulting
matrix is not able to reconstruct the basis of X2.
}
\examples{
x = runif(100)
X1 = cbind(1, x, x^2, x^3, x^4)
X2 = cbind(1, x)
centerMatrix(X1, X2)
}
