% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createBasis}
\alias{createBasis}
\title{De Boors algorithm to find basis functions}
\usage{
createBasis(values, degree, knots)
}
\arguments{
\item{degree}{`unsigned int` Degree of the polynomial between the knots.}

\item{knots}{`arma::vec` Vector of knots. It's the users responsibility to
pass a **SORTED** vector.}

\item{x}{`double` Point to search for position in knots.}
}
\value{
`arma::mat` dense matrix containing the basis functions.
}
\description{
De Boors algorithm to find basis functions
}
\examples{
x = sort(runif(100, 0, 10))
y = 2 * sin(x) + rnorm(100, 0, 0.5)

# Create knots on the space of x:
knots = createKnots(values = x, n_knots = 7, degree = 3)

# Create basis functions for one value:
basisFuns(x = x[30], degree = 3, knots = knots)
}
