% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createKnots}
\alias{createKnots}
\title{Create knots for a specific number, degree and values}
\usage{
createKnots(values, n_knots, degree)
}
\arguments{
\item{values}{`arma::vec` Points to create the basis matrix.}

\item{n_knots}{`unsigned int` Number of innter knots.}

\item{degree}{`unsigned int` polynomial degree of splines.}
}
\value{
`arma::vec` of knots.
}
\description{
This functions takes a vector of points and creates knots used for the
splines depending on the number of knots and degree. This function just
handles equidistant knots.
}
\examples{
x = sort(runif(100, 0, 10))
y = 2 * sin(x) + rnorm(100, 0, 0.5)

# Create knots on the space of x:
knots = createKnots(values = x, n_knots = 7, degree = 3)
}
