% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createSparseBasis}
\alias{createSparseBasis}
\title{Transformation from a vector of input points to sparse matrix of basis}
\usage{
createSparseBasis(values, degree, knots)
}
\arguments{
\item{values}{`arma::vec` Points to create the basis matrix.}

\item{degree}{`unsigned int` polynomial degree of splines.}

\item{n_knots}{`unsigned int` Number of innter knots.}
}
\value{
`arma::sp_mat` sparse matrix of base functions.
}
\description{
This functions takes a vector of points and create a sparse matrix of
basis functions. Each row contains the basis of the corresponding value 
in `values`.
}
\details{
Instead of calculating each row through a helper function we directly 
calculate deboors algorithm here. This is due to the procedure how 
sparse matrices should be allocated and constructed.
}
\examples{
nsim = 100

x = sort(runif(nsim, 0, 10))
y = 2 * sin(x) + rnorm(nsim, 0, 0.5) 
knots = createKnots(values = x, n_knots = 20, degree = 3)

# Create spline basis:
basis = createSparseBasis(values = x, degree = 3, knots = knots)
}
