% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{penaltyMat}
\alias{penaltyMat}
\title{Calculating penalty matrix}
\usage{
penaltyMat(nparams, differences)
}
\arguments{
\item{nparams}{`unsigned int` Number of params which should be penalized.
This also pretend the number of rows and columns.}

\item{differences}{`unsigned int` Number of penalized differences.}
}
\value{
`arma::sp_mat` Sparse penalty matrix used for p splines.
}
\description{
This function calculates the penalty matrix for a given number of 
parameters (`nparams`) and a given number of differences (`differences`).
}
\examples{
pen = penaltyMat(10, 2)
}
