% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createBasis}
\alias{createBasis}
\title{Transformation from a vector of input points to dense matrix of basis}
\usage{
createBasis(values, degree, knots)
}
\arguments{
\item{degree}{[\code{integer(1)]}\cr 
Polynomial degree of splines.}

\item{knots}{[\code{numeric}]\cr 
Vector of knots on which the splines are computed.}
}
\value{
Vector of knots.
}
\description{
This functions takes a vector of points and create a sparse matrix of
basis functions. Each row contains the basis of the corresponding value 
in `values`.
}
\examples{
nsim = 100
x = sort(runif(nsim, 0, 10))
y = 2 * sin(x) + rnorm(nsim, 0, 0.5) 
knots = createKnots(values = x, n_knots = 20, degree = 3)

# Create spline basis:
basis = createBasis(values = x, degree = 3, knots = knots)
}
