% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{createBasis}
\alias{createBasis}
\title{Transformation from a vector of input points to sparse matrix of basis}
\usage{
createBasis(values, degree, knots)
}
\arguments{
\item{values}{`arma::vec` Points to create the basis matrix.}

\item{degree}{`unsigned int` polynomial degree of splines.}

\item{n_knots}{`unsigned int` Number of innter knots.}
}
\value{
`sp_mat` sparse matrix of base functions.
}
\description{
This functions takes a vector of points and create a sparse matrix of
basis functions. Each row contains the basis of the corresponding value 
in `values`.
}
